+++ internal/pkg/runtime/engine/singularity/host_linux.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2022, Sylabs Inc. All rights reserved.
+// Copyright (c) 2022-2026, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -25,17 +25,37 @@ func (e *EngineOperations) PostStartHost(ctx context.Context) (err error) {
 	return nil
 }
 
-// CleanupHost cleans up a SIF FUSE image mount and the temporary directory that
-// holds it. If container creation fails early, in STAGE 1, it will be called
+// CleanupHost cleans up a SIF FUSE image mount and related temporary
+// directories. If container creation fails early, in STAGE 1, it will be called
 // directly from STAGE 1. Otherwise, it will be called from a CLEANUP_HOST
 // process, when the container cleanly exits, or is killed.
 func (e *EngineOperations) CleanupHost(ctx context.Context) (err error) {
-	tmpDir := e.EngineConfig.GetDeleteTempDir()
-	if e.EngineConfig.GetImageFuse() && tmpDir != "" {
+	if !e.EngineConfig.GetImageFuse() {
+		return nil
+	}
+
+	// GetDeleteTempDir being set indicates the rootfs is FUSE mounted in a
+	// temporary directory, which should unmounted and removed. It should have
+	// been cleaned up with a lazy unmount in PostStartHost, but if something
+	// went wrong there, we try again here.
+	if tmpDir := e.EngineConfig.GetDeleteTempDir(); tmpDir != "" {
 		if fs.IsDir(tmpDir) {
+			sylog.Debugf("Cleaning up image FUSE mount temporary directory %s", tmpDir)
 			return cleanFUSETempDir(ctx, e)
 		}
 	}
+
+	// GetDeletePullTempDir being set indicates the underlying image was
+	// implicitly pulled to a temporary directory, due to disabled cache, and
+	// this should be removed.
+	if tmpDir := e.EngineConfig.GetDeletePullTempDir(); tmpDir != "" {
+		sylog.Debugf("Cleaning up image pull temporary directory %s", tmpDir)
+		err := os.RemoveAll(tmpDir)
+		if err != nil {
+			return fmt.Errorf("failed to delete temporary directory %s: %s", tmpDir, err)
+		}
+	}
+
 	return nil
 }
 
@@ -48,7 +68,7 @@ func cleanFUSETempDir(ctx context.Context, e *EngineOperations) error {
 	if tmpDir != "" {
 		err := os.RemoveAll(tmpDir)
 		if err != nil {
-			return fmt.Errorf("failed to delete container image tempDir %s: %s", tmpDir, err)
+			return fmt.Errorf("failed to delete temporary directory %s: %s", tmpDir, err)
 		}
 	}
 	return nil